+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Post v1 a change is further considered breaking if the state file format is inco
 *Minor*
 
 * Moved messages to external files using QuasiQuotes. ([#155](https://github.com/hjwylde/werewolf/issues/155))
+* Moved role names, descriptions and rules to external files using QuasiQuotes. ([#155](https://github.com/hjwylde/werewolf/issues/155))
 
 ### v1.1.1.0